Create EventLogger, move activity logging logic to EventLogger
